Cleanup code.
  * sort usings
  * tabify source
  * move types to separate files
  * fix source analysis issues marked by MonoDevelop
  * no changes in code
